backend/vending_machine/requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jwcrypto==1.4.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.1
psycopg2-binary==2.9.7
pycparser==2.21
pytesseract==0.3.10
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.30.0
rpds-py==0.10.3
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.2
wrapt==1.15.0
